requirements.txt
discord.py
python-dotenv
flask
schedule
openai